+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+## Changed
+
+- `Timespan()` constructor is now `const`.
+
 ## [0.7.1] - 2025-12-10
 
 ### Changed